Update ID handling to use Long instead of Integer
Replaced Integer with Long for IDs across several classes and DTOs to ensure consistency and support larger ID values. Adjusted method signatures, repository interfaces, and query logic accordingly. Removed unnecessary comments in mapping interfaces for better code clarity.

+++ src/main/java/br/gov/agu/nutec/msescala/controller/AvaliadorController.java
@@ -40,14 +40,14 @@ public class AvaliadorController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<AvaliadorResponseDTO> atualizarAvaliador(@PathVariable Integer id, @RequestBody @Valid AvaliadorRequestDTO request, @RequestHeader("Authorization") String token) {
+    public ResponseEntity<AvaliadorResponseDTO> atualizarAvaliador(@PathVariable Long id, @RequestBody @Valid AvaliadorRequestDTO request, @RequestHeader("Authorization") String token) {
         var response = avaliadorService.atualizarAvaliador(id, request, token);
         return ResponseEntity.ok(response);
     }
 
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deletarAvaliador(@PathVariable Integer id) {
+    public ResponseEntity<Void> deletarAvaliador(@PathVariable Long id) {
         avaliadorService.deletarAvaliador(id);
         return ResponseEntity.noContent().build();
     }
